requirements/base.txt
django==5.1.6
python-dotenv==1.0.1
djangorestframework==3.15.0
django-countries==7.6.1
django-phonenumber-field==8.0.0
django-autoslug==1.9.9
phonenumbers==8.13.32
drf-spectacular==0.28.0
pillow==11.1.0
argon2-cffi==23.1.0
djoser==2.3.1
django-taggit==6.1.0
django-filter==24.3
django-celery-email==3.0.0
cloudinary==1.42.0
django-celery-beat==2.7.0
redis==5.2.1
celery==5.4.0
flower==2.0.1
mysqlclient==2.1.1
